added unit tests for other sorts + resolved minor bugs

+++ algorithm/sort/includes/MergeSort.hpp
@@ -2,9 +2,10 @@
 #define MERGESORT_HPP
 #include <vector>
 #include <cstddef>
+#include <cstdint>
 
-void MergeSort(std::vector <size_t> & arr, const size_t &low, const size_t & high);
-void MergeSortIterative(std::vector <size_t> & arr, const size_t &low, const size_t & high);
+void MergeSort(std::vector <size_t> & arr, const int64_t &low, const int64_t & high);
+void MergeSortIterative(std::vector <size_t> & arr, const int64_t &low, const int64_t & high);
 
 #endif /* MERGESORT_HPP */
